// frontend/tina/__generated__/types.ts
export type Maybe<T> = T | null;
export type InputMaybe<T> = Maybe<T>;
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: { input: string; output: string; }
  String: { input: string; output: string; }
  Boolean: { input: boolean; output: boolean; }
  Int: { input: number; output: number; }
  Float: { input: number; output: number; }
  /** References another document, used as a foreign key */
  Reference: { input: any; output: any; }
  JSON: { input: any; output: any; }
};
// ...
export type PageBlocksPricingPlansFeatures = {
  __typename?: 'PageBlocksPricingPlansFeatures';
  feature?: Maybe<Scalars['String']['output']>;
  icon?: Maybe<Scalars['String']['output']>;
};

export type PageBlocksPricingPlans = {
  __typename?: 'PageBlocksPricingPlans';
  name?: Maybe<Scalars['String']['output']>;
  description?: Maybe<Scalars['String']['output']>;
  price?: Maybe<Scalars['Float']['output']>;
  discount?: Maybe<Scalars['String']['output']>;
  block?: Maybe<Scalars['String']['output']>;
  features?: Maybe<Array<Maybe<PageBlocksPricingPlansFeatures>>>;
};

export type PageBlocksPricing = {
  __typename?: 'PageBlocksPricing';
  titleLine1?: Maybe<Scalars['String']['output']>;
  titleLine2?: Maybe<Scalars['String']['output']>;
  discountText?: Maybe<Scalars['String']['output']>;
  plans?: Maybe<Array<Maybe<PageBlocksPricingPlans>>>;
};